requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
# alembic
# psycopg2-binary
# slowapi
# pyjwt
# passlib
# bcrypt==4.0.1
# python-multipart
# email-validator